Connect Masterwork Catalysts to store, etc
Also other misc related performance changes.
