+++ src/components/MasterworkCatalyst/styles.styl
@@ -1,6 +1,7 @@
 .root
   // max-width: 25%
   min-width: 290px
+  max-width: 25%
   flex: 1 0 25%
   margin: 0 !important
   padding: 7px
