--- src/components/NewItem/styles.styl
@@ -1,78 +1,0 @@
-$size = 44px;
-$green = #44bd32;
-
-._root
-  margin: 5px
-
-.root
-  composes: _root
-  display: flex
-
-.imageWrapper
-  position: relative
-
-.obtained
-.dismantled
-.checklisted
-  .fadeOut
-    opacity: .25
-
-.image
-  display: block
-  height: $size
-  width: auto
-  min-width: $size
-  background: rgba(#D2D2D2, 0.1)
-  object-fit: cover
-
-.overlay
-  position: absolute
-  top: 0
-  left: 0
-  width: 100%
-
-.placeholder
-  composes: _root image
-  height: $size
-  background: rgba(#D2D2D2, 0.1)
-
-.tick
-  position: absolute
-  top: -5px
-  right: @top
-  line-height: 1
-  text-align: left
-  background: $green
-  border-radius: 50%
-  font-size: 12px
-  padding: 3px
-
-  .dismantled &
-    background: #7f8fa6
-
-  .checklisted &
-    background: #3c94ff
-
-.extended
-  margin-left: 5px
-  line-height: 1.2
-
-.itemType
-  font-size: .85em
-  opacity: .8
-  margin-top: 2px
-
-.objectiveOverlay
-  position: absolute
-  bottom: 2px
-  left: 2px
-  right: 2px
-  height: 4px
-  background: rgba(black, .5)
-
-.objectiveTrack
-  height: 100%
-  background: $green
-
-.objectivesComplete
-  opacity: .25
